[feat] UserDataStore interface로 캡슐화
