+++ core/datastore/src/main/java/com/moya/funch/datastore/di/DataStoreModule.kt
@@ -2,6 +2,9 @@ package com.moya.funch.datastore.di
 
 import android.content.SharedPreferences
 import com.moya.funch.datastore.PreferenceFactory
+import com.moya.funch.datastore.UserDataStore
+import com.moya.funch.datastore.UserDataStoreImpl
+import dagger.Binds
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -14,4 +17,12 @@ object DataStoreModule {
     @Provides
     @Singleton
     fun provideAppPreferences(factory: PreferenceFactory): SharedPreferences = factory.create()
+
+    @Module
+    @InstallIn(SingletonComponent::class)
+    abstract class Binder {
+        @Binds
+        @Singleton
+        abstract fun bindUserDataStore(userDataStore: UserDataStoreImpl): UserDataStore
+    }
 }
